Made a casual modification to ambiguity_solving.py.
Played around with how the score was computed for each track. Alex (who
worked on the project before I did) already wrote a scoring function
called compute_score (but was not being used).

Removed the score[] and replaced it with compute_score, and precision
and recall are now both 99%, compared to a previous precision of 17% and
recall of 99%. The next step is to look at how the precision/recall are
calculated, to see how meaningful this is.